refactor normalizeURL, return input as is on parse error
